+++ activemq-cpp/src/test/decaf/io/DataInputStreamTest.cpp
@@ -386,7 +386,7 @@ void DataInputStreamTest::test_skipBytes() {
 
         CPPUNIT_ASSERT_NO_THROW_MESSAGE(
             "Should throw an EOFException",
-			skipped = (std::size_t)is->skip( 500000 ) );
+            skipped = (std::size_t)is->skip( 500000 ) );
 
         CPPUNIT_ASSERT_MESSAGE(
             "Skipped should report " +
@@ -592,16 +592,16 @@ void DataInputStreamTest::testUTFDecoding() {
         unsigned char expect[] = {0x00, 0x0B, 0x48, 0x65, 0x6C, 0x6C, 0x6F, 0x20, 0x57, 0x6F, 0x72, 0x6C, 0x64};
         unsigned char input[] = { 0x00, 0x0E ,0xC0, 0x80, 0x0B, 0x48, 0x65, 0x6C, 0x6C, 0x6F, 0x20, 0x57, 0x6F, 0x72, 0x6C, 0x64};
 
-        testHelper( input, sizeof(input)/sizeof(unsigned char),
-                    expect, sizeof(expect)/sizeof(unsigned char) );
+        testHelper( input, (int) sizeof(input) / (int) sizeof(unsigned char),
+                    expect, (int) sizeof(expect) / (int) sizeof(unsigned char) );
     }
 
     // Test data with 2-byte UT8 encoding.
     {
         unsigned char expect[] = {0x00, 0xC2, 0xA9, 0xC3, 0xA6 };
         unsigned char input[] = { 0x00, 0x0A, 0xC0, 0x80, 0xC3, 0x82, 0xC2, 0xA9, 0xC3, 0x83, 0xC2, 0xA6 };
-        testHelper( input, sizeof(input)/sizeof(unsigned char),
-                    expect, sizeof(expect)/sizeof(unsigned char)  );
+        testHelper( input, (int) sizeof(input) / (int) sizeof(unsigned char),
+                    expect, (int) sizeof(expect) / (int) sizeof(unsigned char)  );
     }
 
     // Test data with 1-byte and 2-byte encoding with embedded NULL's.
@@ -609,15 +609,15 @@ void DataInputStreamTest::testUTFDecoding() {
         unsigned char expect[] = {0x00, 0x04, 0xC2, 0xA9, 0xC3, 0x00, 0xA6 };
         unsigned char input[] = { 0x00, 0x0D, 0xC0, 0x80, 0x04, 0xC3, 0x82, 0xC2, 0xA9, 0xC3, 0x83, 0xC0, 0x80, 0xC2, 0xA6 };
 
-        testHelper( input, sizeof(input)/sizeof(unsigned char),
-                    expect, sizeof(expect)/sizeof(unsigned char) );
+        testHelper( input, (int) sizeof(input) / (int) sizeof(unsigned char),
+                    expect, (int) sizeof(expect) / (int) sizeof(unsigned char) );
     }
 
     // Test with bad UTF-8 encoding, missing 2nd byte of two byte value
     {
         unsigned char input[] = { 0x00, 0x0D, 0xC0, 0x80, 0x04, 0xC3, 0x82, 0xC2, 0xC2, 0xC3, 0x83, 0xC0, 0x80, 0xC2, 0xA6 };
 
-        ByteArrayInputStream myStream( input, sizeof(input)/sizeof(unsigned char) );
+        ByteArrayInputStream myStream( input, (int) sizeof(input) / (int) sizeof(unsigned char) );
         DataInputStream reader( &myStream );
 
         CPPUNIT_ASSERT_THROW_MESSAGE(
@@ -630,7 +630,7 @@ void DataInputStreamTest::testUTFDecoding() {
     {
         unsigned char input[] = { 0x00, 0x0D, 0xC0, 0x80, 0x04, 0xC3, 0x82, 0xC2, 0xC2, 0xC3, 0x83, 0xC0, 0x80, 0xC2, 0xA6 };
 
-        ByteArrayInputStream myStream( input, sizeof(input)/sizeof(unsigned char) );
+        ByteArrayInputStream myStream( input, (int) sizeof(input) / (int) sizeof(unsigned char) );
         DataInputStream reader( &myStream );
 
         CPPUNIT_ASSERT_THROW_MESSAGE(
@@ -642,7 +642,7 @@ void DataInputStreamTest::testUTFDecoding() {
     // Test data with value greater than 255 in 2-byte encoding.
     {
         unsigned char input[] = {0x00, 0x04, 0xC8, 0xA9, 0xC3, 0xA6};
-        ByteArrayInputStream myStream( input, sizeof(input)/sizeof(unsigned char) );
+        ByteArrayInputStream myStream( input, (int) sizeof(input) / (int) sizeof(unsigned char) );
         DataInputStream reader( &myStream );
 
         CPPUNIT_ASSERT_THROW_MESSAGE(
@@ -654,7 +654,7 @@ void DataInputStreamTest::testUTFDecoding() {
     // Test data with value greater than 255 in 3-byte encoding.
     {
         unsigned char input[] = {0x00, 0x05, 0xE8, 0xA8, 0xA9, 0xC3, 0xA6};
-        ByteArrayInputStream myStream( input, sizeof(input)/sizeof(unsigned char) );
+        ByteArrayInputStream myStream( input, (int) sizeof(input) / (int) sizeof(unsigned char) );
         DataInputStream reader( &myStream );
 
         CPPUNIT_ASSERT_THROW_MESSAGE(
@@ -666,7 +666,7 @@ void DataInputStreamTest::testUTFDecoding() {
     // Test with three byte encode that's missing a last byte.
     {
         unsigned char input[] = {0x00, 0x02, 0xE8, 0xA8};
-        ByteArrayInputStream myStream( input, sizeof(input)/sizeof(unsigned char) );
+        ByteArrayInputStream myStream( input, (int) sizeof(input) / (int) sizeof(unsigned char) );
         DataInputStream reader( &myStream );
 
         CPPUNIT_ASSERT_THROW_MESSAGE(
